Select OPD units based on BUNIT header keyword
poppy.Instrument will no longer set opdunits='micron' by default,
allowing the units of OPD to be set from the FITS header. The
code in FITSOpticalElement to handle this has been changed,
and the signature of FITSOpticalElement.__init__ changed so that
the FITS keyword inspection code will not be skipped.