Remove wavelength limit from instrument.py
I'm working with @mperrin on PSF subtraction and came across a scenario where I wanted to generate a data cube with more than 100 slices but was hit with an error -- `ValueError: Maximum number of wavelengths exceeded. Cannot be more than 100`. 

I brought it up to Marshall and he couldn't remember why this limit existed, so I'm proposing a pull request that removes it to see what `poppy`'s other developers think.