Moved FITSFPMElement, additional unit handling for displaying, changed FresnelWavefront__Imul__
I moved FITSFPMElement to be in fresnel.py instead of poppy_core.py because it is only designed for use by a FresnelWavefront.

Slightly changed apply_fits_fpm_fftmft to use pixelscale_lamD directly.

Changed FresnelWavefront __imul__ method to use apply_fits_fpm_fftmft when necessary, instead of calling the method within the propagate() method.

Added additional unit handling in the display method in OpticalElement class. This will allow for FITSFPMElements to be displayed with the extents in units of arcseconds as they should be.